+++ prototype_ui/assets/dynamic/js/pages/reports/404.js
@@ -1,0 +1,6 @@
+export const Page404 = {
+    mount(params) {
+        const container = document.getElementById("page-404");
+        container.innerHTML = ``;
+    }
+}