+++ prototype_ui/assets/dynamic/js/router/index.js
@@ -1,12 +1,20 @@
 import {Dispatcher} from "../composables/useDispatcher.js";
 
 export function init(app) {
+    app.routing.add("sign_in", () => {
+        app.panelManager.mountScene("sign_in");
+    });
+
+    app.routing.add("sign_out", () => {
+        app.panelManager.mountScene("sign_out");
+    });
+
     app.routing.add("dashboard", () => {
-        app.panelManager.mountPanels();
+        app.panelManager.mountScene("dashboard");
     });
 
     app.routing.add("add_entry", () => {
-        app.root.innerHTML = "<h1>Add entry</h1>";
+        app.panelManager.mountScene("add_entry")
     });
 
     app.routing.add("404", () => {